--- heimdall-core/src/main/java/br/com/conductor/heimdall/core/interceptor/RattingHeimdallInterceptor.java
@@ -1,34 +1,0 @@
-package br.com.conductor.heimdall.core.interceptor;
-
-import br.com.conductor.heimdall.core.dto.interceptor.RateLimitDTO;
-import br.com.conductor.heimdall.core.enums.TypeExecutionPoint;
-import br.com.conductor.heimdall.core.interfaces.HeimdallInterceptor;
-
-import java.util.HashMap;
-
-/**
- * Implementation of the HeimdallInterceptor to type Ratting.
- *
- * @author <a href="https://dijalmasilva.github.io" target="_blank">Dijalma Silva</a>
- */
-public class RattingHeimdallInterceptor implements HeimdallInterceptor {
-
-    @Override
-    public String getFile(TypeExecutionPoint typeExecutionPoint) {
-        return "ratting.mustache";
-    }
-
-    @Override
-    public boolean validateTemplate(Object objectCustom) {
-        return false;
-    }
-
-    @Override
-    public HashMap<String, Object> buildParameters(Object objectCustom, HashMap<String, Object> parameters) {
-        RateLimitDTO rateLimitDTO = (RateLimitDTO) objectCustom;
-        parameters.put("calls", rateLimitDTO.getCalls());
-        parameters.put("interval", rateLimitDTO.getInterval().name());
-
-        return parameters;
-    }
-}
