+++ heimdall-core/src/main/java/br/com/conductor/heimdall/core/interceptor/impl/LogHeimdallInterceptor.java
@@ -1,5 +1,4 @@
-
-package br.com.conductor.heimdall.core.enums;
+package br.com.conductor.heimdall.core.interceptor.impl;
 
 /*-
  * =========================LICENSE_START==================================
@@ -10,9 +9,9 @@ package br.com.conductor.heimdall.core.enums;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -22,35 +21,35 @@ package br.com.conductor.heimdall.core.enums;
  */
 
 import br.com.conductor.heimdall.core.entity.Interceptor;
+import br.com.conductor.heimdall.core.enums.TypeExecutionPoint;
 import br.com.conductor.heimdall.core.interceptor.HeimdallInterceptor;
-import br.com.conductor.heimdall.core.interceptor.impl.*;
-import lombok.Getter;
+
+import java.util.HashMap;
 
 /**
- * Enum that hold the types of {@link Interceptor}.
+ * Implementation of the HeimdallInterceptor to type Log.
  *
- * @author Filipe Germano
  * @author <a href="https://dijalmasilva.github.io" target="_blank">Dijalma Silva</a>
- *
  */
-@Getter
-public enum TypeInterceptor {
+public class LogHeimdallInterceptor implements HeimdallInterceptor {
 
-	LOG(new LogHeimdallInterceptor()),
-	MOCK(new MockHeimdallInterceptor()),
-	RATTING(new RattingHeimdallInterceptor()),
-	ACCESS_TOKEN(new AccessTokenHeimdallInterceptor()),
-	CLIENT_ID(new ClientIdHeimdallInterceptor()),
-	CUSTOM(new CustomHeimdallInterceptor()),
-	MIDDLEWARE(new MiddlewareHeimdallInterceptor()),
-	OAUTH(new OAuthHeimdallInterceptor()),
-	BLACKLIST(new BlacklistHeimdallInterceptor()),
-	WHITELIST(new WhitelistHeimdallInterceptor());
+    @Override
+    public String getFile(TypeExecutionPoint typeExecutionPoint) {
 
-	private HeimdallInterceptor heimdallInterceptor;
+        if (typeExecutionPoint.equals(TypeExecutionPoint.FIRST)) {
+            return "request_log.mustache";
+        }
 
-	TypeInterceptor(HeimdallInterceptor heimdallInterceptor) {
-		this.heimdallInterceptor = heimdallInterceptor;
-	}
+        return "response_log.mustache";
+    }
 
+    @Override
+    public Object parseContent(String content) {
+        return content;
+    }
+
+    @Override
+    public HashMap<String, Object> buildParameters(Object objectCustom, HashMap<String, Object> parameters, Interceptor interceptor) {
+        return parameters;
+    }
 }
