+++ heimdall-core/src/main/java/br/com/conductor/heimdall/core/service/InterceptorService.java
@@ -227,148 +227,7 @@ public class InterceptorService {
     }
 
     private Object validateTemplate(TypeInterceptor type, String content) {
-
-        boolean valid = false;
-        switch (type) {
-            case ACCESS_TOKEN:
-                try {
-
-                    AccessTokenClientIdDTO accessTokenDTO = JsonUtils.convertJsonToObject(content, AccessTokenClientIdDTO.class);
-                    if (Objeto.notBlank(accessTokenDTO)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_ACCESS_TOKEN);
-                }
-                break;
-            case CLIENT_ID:
-                try {
-
-                    AccessTokenClientIdDTO clientIdDTO = JsonUtils.convertJsonToObject(content, AccessTokenClientIdDTO.class);
-                    if (Objeto.notBlank(clientIdDTO)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_ACCESS_TOKEN);
-                }
-                break;
-            case MOCK:
-                try {
-
-                    MockDTO mockDTO = JsonUtils.convertJsonToObject(content, MockDTO.class);
-                    if (Objeto.notBlank(mockDTO)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_MOCK);
-                }
-                break;
-            case RATTING:
-                try {
-
-                    RateLimitDTO rateLimitDTO = JsonUtils.convertJsonToObject(content, RateLimitDTO.class);
-                    if (Objeto.notBlank(rateLimitDTO)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_RATTING);
-                }
-                break;
-            case CUSTOM:
-                try {
-
-                    if (Objeto.notBlank(content)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                }
-                break;
-            case MIDDLEWARE:
-                try {
-
-                    if (Objeto.notBlank(content)) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                }
-                break;
-            case OAUTH:
-                try {
-                    OAuthDTO oAuthDTO = JsonUtils.convertJsonToObject(content, OAuthDTO.class);
-
-                    if (Objeto.isBlank(oAuthDTO.getTypeOAuth())) {
-                        ExceptionMessage.TYPE_OAUTH_NOT_FOUND.raise(type.name(), ExceptionMessage.TYPE_OAUTH_NOT_FOUND.getMessage());
-                    }
-
-                    if (oAuthDTO.getTypeOAuth().equals(TypeOAuth.AUTHORIZE.getTypeOAuth())) {
-                        if (Objeto.isBlank(oAuthDTO.getProviderId()) || oAuthDTO.getProviderId() == 0) {
-                            ExceptionMessage.PROVIDER_NOT_FOUND.raise(type.name(), ExceptionMessage.PROVIDER_NOT_FOUND.getMessage());
-                        }
-                    }
-
-                    if (oAuthDTO.getTypeOAuth().equals(TypeOAuth.GENERATE.getTypeOAuth()) || oAuthDTO.getTypeOAuth().equals(TypeOAuth.VALIDATE.getTypeOAuth())) {
-                        if (Objeto.isBlank(oAuthDTO.getPrivateKey())) {
-                            ExceptionMessage.PRIVATE_KEY_NOT_FOUND.raise(type.name(), ExceptionMessage.PRIVATE_KEY_NOT_FOUND.getMessage());
-                        }
-                    }
-
-                    valid = true;
-
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_OAUTH);
-                }
-                break;
-            case BLACKLIST:
-                try {
-
-                    if (Objeto.notBlank(JsonUtils.convertJsonToObject(content, IpsDTO.class))) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_BLOCK_IPS);
-                }
-                break;
-            case WHITELIST:
-                try {
-
-                    if (Objeto.notBlank(JsonUtils.convertJsonToObject(content, IpsDTO.class))) {
-
-                        valid = true;
-                    }
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_BLOCK_IPS);
-                }
-                break;
-            default:
-                break;
-        }
-
-        return valid;
+        return type.getHeimdallInterceptor().parseContent(content);
     }
 
     /**
