+++ heimdall-gateway/src/main/java/br/com/conductor/heimdall/gateway/service/InterceptorFileService.java
@@ -118,6 +118,7 @@ public class InterceptorFileService {
             parameters.put("pathsNotAllowed", pathsNotAllowed(interceptor));
             parameters.put("lifeCycle", interceptor.getLifeCycle().name());
             parameters.put("name", StringUtils.concatCamelCase(interceptor.getLifeCycle().name(), interceptor.getType().name(), interceptor.getExecutionPoint().getFilterType(), interceptor.getId().toString()));
+            parameters.put("zuulFilterRoot", zuulFilterRoot);
             if (Objeto.notBlank(interceptor.getOperation())) {
 
                 parameters.put("method", interceptor.getOperation().getMethod().name());
@@ -281,15 +282,11 @@ public class InterceptorFileService {
 
     private HashMap<String, Object> buildCustom(HashMap<String, Object> parameters, Interceptor interceptor) {
 
-        Object objectCustom = validateTemplate(interceptor.getType(), interceptor.getContent());
+        Object objectCustom = interceptor.getType().getHeimdallInterceptor().parseContent(interceptor.getContent());
+
         if (Objeto.notBlank(objectCustom)) {
 
             if (objectCustom instanceof AccessTokenClientIdDTO) {
-
-                AccessTokenClientIdDTO accessTokenClientIdDTO = (AccessTokenClientIdDTO) objectCustom;
-                parameters.put("name", accessTokenClientIdDTO.getName());
-                parameters.put("location", accessTokenClientIdDTO.getLocation());
-                parameters.put("apiId", interceptor.getApi().getId());
 
                 if (TypeInterceptor.ACCESS_TOKEN.equals(interceptor.getType())) {
 
@@ -328,155 +325,11 @@ public class InterceptorFileService {
 
             }
 
-            if (objectCustom instanceof MockDTO) {
-
-                MockDTO mockDTO = (MockDTO) objectCustom;
-                parameters.put("status", mockDTO.getStatus());
-                parameters.put("body", mockDTO.getBody());
-
-            }
-
-            if (objectCustom instanceof RateLimitDTO) {
-
-                RateLimitDTO rateLimitDTO = (RateLimitDTO) objectCustom;
-                parameters.put("calls", rateLimitDTO.getCalls());
-                parameters.put("interval", rateLimitDTO.getInterval().name());
-            }
-
-            if (objectCustom instanceof OAuthDTO) {
-
-                OAuthDTO oAuthDTO = (OAuthDTO) objectCustom;
-
-                parameters.put("providerId", Objeto.isBlank(oAuthDTO.getProviderId()) ? 0L : oAuthDTO.getProviderId());
-                parameters.put("timeAccessToken", Objeto.isBlank(oAuthDTO.getTimeAccessToken()) ? 20 : oAuthDTO.getTimeAccessToken());
-                parameters.put("timeRefreshToken", Objeto.isBlank(oAuthDTO.getTimeRefreshToken()) ? 1800 : oAuthDTO.getTimeRefreshToken());
-                parameters.put("typeOAuth", oAuthDTO.getTypeOAuth());
-                parameters.put("privateKey", oAuthDTO.getPrivateKey());
-            }
-
-            if (objectCustom instanceof IpsDTO){
-
-                IpsDTO ipsDTO = (IpsDTO) objectCustom;
-                parameters.put("ips", ipsDTO.getIps());
-            }
-
-            if (objectCustom instanceof String) {
-
-                parameters.put("content", objectCustom);
-            }
-        }
-
-        if (TypeInterceptor.MIDDLEWARE.equals(interceptor.getType())) {
-
-            String api = interceptor.getOperation().getResource().getApi().getId().toString();
-            String pathReferences = String.join("/", zuulFilterRoot, MIDDLEWARE_API_ROOT, api, MIDDLEWARE_ROOT);
-            parameters.put("pathMiddleware", pathReferences);
+            parameters = interceptor.getType().getHeimdallInterceptor().buildParameters(objectCustom, parameters, interceptor);
         }
 
         return parameters;
     }
-
-    private Object validateTemplate(TypeInterceptor type, String content) {
-
-        Object response = null;
-        switch (type) {
-            case ACCESS_TOKEN:
-                try {
-
-                    AccessTokenClientIdDTO accessTokenDTO = JsonUtils.convertJsonToObject(content, AccessTokenClientIdDTO.class);
-                    response = accessTokenDTO;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_ACCESS_TOKEN);
-                }
-                break;
-            case CLIENT_ID:
-                try {
-
-                    AccessTokenClientIdDTO clientIdDTO = JsonUtils.convertJsonToObject(content, AccessTokenClientIdDTO.class);
-                    response = clientIdDTO;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_ACCESS_TOKEN);
-                }
-                break;
-            case MOCK:
-                try {
-
-                    MockDTO mockDTO = JsonUtils.convertJsonToObject(content, MockDTO.class);
-                    response = mockDTO;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_MOCK);
-                }
-                break;
-            case RATTING:
-                try {
-
-                    RateLimitDTO rateLimitDTO = JsonUtils.convertJsonToObject(content, RateLimitDTO.class);
-                    response = rateLimitDTO;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_RATTING);
-                }
-                break;
-            case CUSTOM:
-                try {
-
-                    response = content;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                }
-                break;
-            case MIDDLEWARE:
-                try {
-
-                    response = content;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                }
-                break;
-            case OAUTH:
-                try {
-
-                    OAuthDTO oAuthDTO = JsonUtils.convertJsonToObject(content, OAuthDTO.class);
-                    response = oAuthDTO;
-                } catch (Exception e) {
-
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_OAUTH);
-                }
-                break;
-            case BLACKLIST:
-                try {
-                    response = JsonUtils.convertJsonToObject(content, IpsDTO.class);
-                } catch (IOException e) {
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_BLOCK_IPS);
-                }
-                break;
-            case WHITELIST:
-                try {
-                    response = JsonUtils.convertJsonToObject(content, IpsDTO.class);
-                } catch (IOException e) {
-                    log.error(e.getMessage(), e);
-                    ExceptionMessage.INTERCEPTOR_INVALID_CONTENT.raise(type.name(), TemplateUtils.TEMPLATE_BLOCK_IPS);
-                }
-                break;
-            default:
-                break;
-        }
-
-        return response;
-    }
-
-
 
     private void generateFileInterceptor(Interceptor interceptor, String template, HashMap<String, Object> parameters) {
